+++ python/eyes_images/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [vNext]
+### Fixed
+- Permit compatible values of enums in Configuration objects [Trello 3188](https://trello.com/c/VaZniTQ4)
+
 ## [5.17.0] - 2023-03-31
 ### Added
 - Galaxy S22, Galaxy Tab S7, Huawei Mate 50 Pro, Huawei Matepad 11 and Sony Xperia 10 II to the enum of supported emulators [Trello 3187](https://trello.com/c/mnXe35i8)
@@ -10,7 +14,7 @@
 ## [5.16.0] - 2023-03-08
 ### Added
 - Support webview targets with appium [Trello 3162](https://trello.com/c/77BDT6t6)
-## Fixed
+### Fixed
 - Add missing set_configuration/get_configuration methods and configuration property to images.Eyes [Trello 3155](https://trello.com/c/QZz64qr3)
 ### Updated
 - Drop obsolete eyes-core and eyes-common dependencies [Trello 3165](https://trello.com/c/aLfsFODx)